Replace `zonal_mean_2d_plot.py` - Update function in `zonal_mean_2d_stratosphere_ploy.py`
